pyproject.toml
[tool.poetry]
name = "mydrift"
version = "0.1.0"
description = ""
authors = ["windysavage"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
pytest = "^8.3.5"
ruff = "0.11.2"
qdrant-client = "^1.13.0"
torch = "2.6.0"
sentence-transformers = "^4.0.1"
ollama = "^0.4.7"
attrs = "^25.3.0"
python-dotenv = "^1.1.0"
pydantic = "^2.11.1"
pydantic-settings = "^2.8.1"
fastapi = "^0.115.12"
uvicorn = "^0.34.0"

[tool.poetry.group.jupyter.dependencies]
ipykernel = ">=6.29,<7.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
